+++ test/sim_info_plugin_test.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sim_info_plugin/sim_info_plugin.dart';
 import 'package:sim_info_plugin/sim_info_plugin_platform_interface.dart';
@@ -10,6 +12,13 @@ class MockSimInfoPluginPlatform
 
   @override
   Future<List<dynamic>?> getSimCardsDirect() => Future.value([{'number': '123'}]);
+
+  @override
+  Stream<List<dynamic>> getSimCardsStream() {
+    return Stream<List<dynamic>>.value([
+      {'number': '123'},
+    ]);
+  }
 }
 
 void main() {
@@ -25,5 +34,15 @@ void main() {
     SimInfoPluginPlatform.instance = fakePlatform;
 
     expect(await simInfoPlugin.getSimCardsDirect(), [{'number': '123'}]);
+  });
+
+  test('getSimCardsStream', () async {
+    SimInfoPlugin simInfoPlugin = SimInfoPlugin();
+    MockSimInfoPluginPlatform fakePlatform = MockSimInfoPluginPlatform();
+    SimInfoPluginPlatform.instance = fakePlatform;
+
+    expect(await simInfoPlugin.getSimCardsStream().first, [
+      {'number': '123'},
+    ]);
   });
 }